+++ Agendamento/src/main/java/br/com/NexusSolutions/Agendamento/ServicoClient/ResponseDTO.java
@@ -1,0 +1,8 @@
+package br.com.NexusSolutions.Agendamento.ServicoClient;
+
+import lombok.Data;
+
+@Data
+public class ResponseDTO {
+    private Long id;
+}
