+++ Agendamento/src/main/java/br/com/NexusSolutions/Agendamento/ServicoClient/ServicoClient.java
@@ -1,0 +1,5 @@
+package br.com.NexusSolutions.Agendamento.ServicoClient;
+
+
+public class ServicoClient {
+}
