+++ pom.xml
@@ -11,8 +11,13 @@
     <groupId>br.com.NexusSolutions</groupId>
     <artifactId>WebProjectClinicaMedica</artifactId>
     <version>0.0.1-SNAPSHOT</version>
+    <packaging>pom</packaging>
     <name>WebProjectClinicaMedica</name>
     <description>Demo project for Spring Boot</description>
+    <modules>
+        <module>Administrativo</module>
+        <module>Agendamento</module>
+    </modules>
     <properties>
         <java.version>17</java.version>
     </properties>
