--- src/main/java/br/com/NexusSolutions/WebProjectClinicaMedica/Exception/ResponseException.java
@@ -1,15 +1,0 @@
-package br.com.NexusSolutions.WebProjectClinicaMedica.Exception;
-
-import lombok.Builder;
-import lombok.Data;
-
-import java.util.List;
-
-@Builder
-@Data
-public class ResponseException {
-    private String originalError;
-    private int httpCode;
-    private String mensagem;
-    private List<String> campos;
-}
